backend/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.4.0
bcrypt>=4.0.0
python-multipart>=0.0.6
sqlalchemy[asyncio]>=2.0.0
psycopg[binary]>=3.2.0
asyncpg>=0.30.0
alembic>=1.16.0 